requirements.txt
hazm==0.7.0
wordcloud==1.8.2.2
arabic-reshaper==2.1.3
python-bidi==0.4.2
loguru==0.6.0